Backend/python-ai-service/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
python-multipart>=0.0.9
pdfplumber>=0.10.0
spacy>=3.7.0
python-docx>=1.1.0
scikit-learn>=1.4.0
groq>=0.5.0
python-dotenv>=1.0.1
requests>=2.31.0
